+++ src/main/java/com/hibit2/hibit2/post/controller/PostController.java
@@ -2,12 +2,21 @@ package com.hibit2.hibit2.post.controller;
 
 
 
+import com.hibit2.hibit2.auth.dto.LoginMember;
+import com.hibit2.hibit2.auth.presentation.AuthenticationPrincipal;
+import com.hibit2.hibit2.member.domain.Member;
+import com.hibit2.hibit2.member.repository.MemberRepository;
 import com.hibit2.hibit2.post.domain.Post;
 import com.hibit2.hibit2.post.dto.PostListDto;
 import com.hibit2.hibit2.post.dto.PostResponseDto;
 import com.hibit2.hibit2.post.dto.PostSaveDto;
 import com.hibit2.hibit2.post.dto.PostUpdateDto;
+import com.hibit2.hibit2.post.repository.PostRepository;
 import com.hibit2.hibit2.post.service.PostService;
+import com.hibit2.hibit2.postHistory.domain.postHistory;
+import com.hibit2.hibit2.postHistory.repository.postHistoryRepository;
+import com.hibit2.hibit2.user.domain.Users;
+import com.hibit2.hibit2.user.repository.UsersRepository;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
@@ -33,11 +42,13 @@ import java.util.List;
 @RequestMapping("/post")
 public class PostController {
     private final PostService postService;
+    private final postHistoryRepository postHistoryRepository;
+    private final MemberRepository memberRepository;
+    private final PostRepository postRepository;
 
 
-    @PostMapping("/write")
-    //@Secured({"ROLE_USER"})
-    @Operation(summary = "post/write", description = "매칭 게시글 작성")
+    @PostMapping("/write/{member_idx}")
+    @Operation(summary = "post/write/{member_idx}", description = "매칭 게시글 작성")
     @Parameters({@Parameter(name = "title", description = "제목", example = "디뮤지엄 전시 보러가요"),
                 @Parameter(name = "content", description = "내용", example =  "본문내용"),
                 @Parameter(name = "number", description = "관람 인원", example =  "3"),
@@ -50,11 +61,9 @@ public class PostController {
                @Parameter(name = "mainimg", description = "대표이미지url", example ="hibitbucket"),
                @Parameter(name = "subimg", description = "서브이미지URL 리스트", example ="hibitbucket")
     })
-    public ResponseEntity<Post> save(@RequestBody PostSaveDto requestDto){
-        //Users user = (Users) authentication.getPrincipal();
-        Post post = postService.save(requestDto);
+    public ResponseEntity<Post> save(@RequestBody PostSaveDto requestDto,@PathVariable Long member_idx){
+        Post post = postService.save(requestDto, member_idx);
         return ResponseEntity.status(HttpStatus.CREATED).body(post);
-
     }
     //기본 게시글 리스트
     @GetMapping("/list/allposts/{pageParam}")
@@ -111,7 +120,7 @@ public class PostController {
     }
 
     // 검색
-    @GetMapping("/search/{pageParam}")
+    @GetMapping("/list/search/{pageParam}")
     @Operation(summary = "post/list/search/1", description = "검색 결과")
     public ResponseEntity<List<PostListDto>> findByTitleOrExhibition(@PathVariable int pageParam, @RequestParam String keyword) {
         char flag = 'D';
@@ -131,25 +140,62 @@ public class PostController {
 
     @PutMapping("/{post_idx}")
     @Operation(summary = "post/{post_idx}",description = "매칭 글 수정")
-    public ResponseEntity<Post> update(@PathVariable int post_idx, @RequestBody PostUpdateDto requestDto){
-        Post post = postService.update(post_idx, requestDto);
-        return ResponseEntity.status(HttpStatus.CREATED).body(post);
+    public ResponseEntity<Post> update(@Parameter(hidden = true) @AuthenticationPrincipal final LoginMember loginMember,@PathVariable int post_idx, @RequestBody PostUpdateDto requestDto){
+        Post entity= postRepository.findById(post_idx).orElseThrow(()-> new IllegalArgumentException("해당 게시글이 없습니다. id="+post_idx));
+        if (loginMember.getId() == entity.getMember().getId()) {
+            Post post = postService.update(post_idx, requestDto);
+            return ResponseEntity.status(HttpStatus.CREATED).body(post);
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
     }
 
     @DeleteMapping("/{post_idx}")
     @Operation(summary = "post/{post_idx}", description = "매칭 글 삭제")
-    public ResponseEntity<Void> delete(@PathVariable int post_idx){
-        postService.delete(post_idx);
+    public ResponseEntity<Void> delete(@Parameter(hidden = true) @AuthenticationPrincipal final LoginMember loginMember, @PathVariable int post_idx){
+        Post entity= postRepository.findById(post_idx).orElseThrow(()-> new IllegalArgumentException("해당 게시글이 없습니다. id="+post_idx));
+        if (loginMember.getId() == entity.getMember().getId()){
+            postService.delete(post_idx);
+        }
+        else{
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         return ResponseEntity.ok().build();
 
     }
+
     //게시글 좋아요
-    @GetMapping("/{post_idx}/like")
-    @Operation(summary = "post/{post_idx}/like", description = "좋아요 누르기, 테스트할려면 유저 signup에서 b로 회원가입하기")
-    public ResponseEntity<Post> likeComment(@PathVariable int post_idx){
-        String user_id = "b"; //나중에는 현재 로그인한 유저의 id 찾아오기
-        Post post = postService.likePost(post_idx, user_id);
-        return ResponseEntity.ok(post);
+    @GetMapping("/{post_idx}/{member_idx}/like")
+    @Operation(summary = "post/{post_idx}/{member_idx}/like", description = "좋아요 누르기, 테스트할려면 유저 signup에서 b로 회원가입하기")
+    public ResponseEntity<PostResponseDto> likeComment(@PathVariable int post_idx, @PathVariable Long member_idx){
+        Post post = postService.likePost(post_idx, member_idx);
+        PostResponseDto postResponseDto = new PostResponseDto(post);
+        return ResponseEntity.ok(postResponseDto);
     }
+
+    //게시글 상태 변경
+    @PutMapping("/{post_idx}/complete")
+    @Operation(summary = "/post/1/complete", description = "게시글 모집 완료")
+    public ResponseEntity<String> completePost(@PathVariable int post_idx) {
+        postHistory postHistory = postHistoryRepository.findByPostIdx(post_idx);
+        if (postHistory.getOkNum() == 0) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("매칭이 진행되지 않았습니다. 매칭 진행 이후 모집 완료를 눌러주세요.");
+        }
+        postService.completePost(post_idx);
+
+        return ResponseEntity.ok().build();
+    }
+
+    //게시글 취소 변경
+    @PutMapping("/{post_idx}/cancle")
+    @Operation(summary = "/post/1/cancle", description = "게시글 모집 완료")
+    public ResponseEntity<String> canclePost(@PathVariable int post_idx) {
+        postService.canclePost(post_idx);
+        return ResponseEntity.ok().build();
+    }
+
+
+
 
 }
